+++ CHANGELOG.md
@@ -6,6 +6,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
 
+## [UNRELEASED]
+### Fixed/Changed
+ - Bug: if locale matcher found an exact match, extensions would not be applied to the resulting tag.
+
 ## [0.14.2]
 ### Fixed/Changed
  - Synchronize enhanced language distance rules with upstream. Minor change in language distance between `en-GB` and `en-US`.
@@ -118,6 +122,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 #### Fixed/Changed
 - Converted `CLDR` to an interface
 
+[UNRELEASED]: https://github.com/phensley/cldr-engine/compare/v0.14.2...master
 [0.14.2]: https://github.com/phensley/cldr-engine/compare/v0.14.1...v0.14.2
 [0.14.1]: https://github.com/phensley/cldr-engine/compare/v0.13.7...v0.14.1
 [0.13.7]: https://github.com/phensley/cldr-engine/compare/v0.13.6...v0.13.7